backEnd/app/requirement.txt
astroid==2.4.2
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.2
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
PyGreSQL==5.2
pylint==2.5.3
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
